moved session stuff to package
