+++ src/main/java/com/marginallyclever/robotoverlord/parameters/RemoteEntity.java
@@ -1,11 +1,10 @@
 package com.marginallyclever.robotoverlord.parameters;
 
-import com.marginallyclever.communications.SessionLayer;
-import com.marginallyclever.communications.SessionLayerEvent;
-import com.marginallyclever.communications.SessionLayerListener;
-import com.marginallyclever.communications.SessionLayerManager;
+import com.marginallyclever.communications.session.SessionLayer;
+import com.marginallyclever.communications.session.SessionLayerEvent;
+import com.marginallyclever.communications.session.SessionLayerListener;
+import com.marginallyclever.communications.session.SessionLayerManager;
 import com.marginallyclever.convenience.StringHelper;
-import com.marginallyclever.convenience.log.Log;
 import com.marginallyclever.robotoverlord.Entity;
 import com.marginallyclever.robotoverlord.RobotOverlord;
 import org.slf4j.Logger;
